Mas i1847 conditionalput (#86)
Add support for `if_not_modified` option, which should work as with riak-erlang-client (and use the X-Riak-If-Not-Modified) header in Riak 3.0.13.

Add support for using `if_none_match`, and `if_match`.  However, if possible, `if_not_modified` should always in preference to `if_match` as `if_match` requires re-calculation of the etag, using knowledge that should be internal to Riak.